Move the installation function to the lib dir pla-92
