+++ lib/install.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 import {fileURLToPath} from 'url'
 import path from 'path'
 import fs from 'fs'
